+++ src/sqlitely/components.py
@@ -8,7 +8,7 @@ Released under the MIT License.
 
 @author      Erki Suurjaak
 @created     21.08.2019
-@modified    30.09.2024
+@modified    01.10.2024
 ------------------------------------------------------------------------------
 """
 import base64
@@ -3675,7 +3675,7 @@ class SchemaObjectPage(wx.Panel):
 
         check_rowid = self._ctrls["without"] = wx.CheckBox(panel, label="WITHOUT &ROWID")
         check_strict = None
-        if self._db.has_feature("strict") \
+        if self._db.has_support("strict") \
         or any(x.get("strict") for x in util.getval(self._item, "meta", "options") or []):
             check_strict = self._ctrls["strict"]  = wx.CheckBox(panel, label="STRICT")
         check_exists = self._ctrls["exists"]  = wx.CheckBox(panel, label="IF N&OT EXISTS")
@@ -3694,7 +3694,7 @@ class SchemaObjectPage(wx.Panel):
         check_exists.ToolTip = "Add 'IF NOT EXISTS' to CREATE SQL statement.\n\n" \
                                "Does not affect creation within this database,\n" \
                                "merely becomes part of schema SQL."
-        if check_strict and not self._db.has_feature("strict"):
+        if check_strict and not self._db.has_support("strict"):
             check_strict.Disable()
             check_strict._toggle = "skip"
 
@@ -4300,7 +4300,7 @@ class SchemaObjectPage(wx.Panel):
 
         self._EmptyControl(self._panel_columns)
         p1, p2 = self._panel_splitter.Children
-        if self._db.has_feature("view_columns") and (items or self._editmode):
+        if self._db.has_support("view_columns") and (items or self._editmode):
             self._panel_splitter.SplitHorizontally(p1, p2, self._panel_splitter.MinimumPaneSize)
             grid.AppendRows(len(items))
             for i, coldata in enumerate(items):
@@ -4866,7 +4866,7 @@ class SchemaObjectPage(wx.Panel):
         if can_simple and droppedcols:
             can_simple = False # There are deleted columns
         if can_simple and any(colmap2[x]["name"] != colmap1[x]["name"] for x in colmap1):
-            can_simple = self._db.has_feature("rename_column") # There are renamed columns
+            can_simple = self._db.has_support("rename_column") # There are renamed columns
         if can_simple:
             if any(x["__id__"] not in colmap1 and cols2[i+1]["__id__"] in colmap1
                    for i, x in enumerate(cols2[:-1])):
@@ -4894,7 +4894,7 @@ class SchemaObjectPage(wx.Panel):
                              and not ("fk" in c2 and self._fks_on and default != "NULL")
                 if not can_simple: break # for c2
         if can_simple and old["name"] != new["name"] \
-        and not self._db.has_feature("full_rename_table"):
+        and not self._db.has_support("full_rename_table"):
             if util.lceq(old["name"], new["name"]): # Case changed
                 can_simple = False
             else:
@@ -6249,7 +6249,7 @@ class SchemaObjectPage(wx.Panel):
             # Show or hide view/trigger columns section where not relevant
             if "view" == self._category:
                 splitter, (p1, p2) = self._panel_splitter, self._panel_splitter.Children
-                if self._db.has_feature("view_columns") \
+                if self._db.has_support("view_columns") \
                 and (self._item["meta"].get("columns") or self._editmode):
                     splitter.SplitHorizontally(p1, p2, splitter.MinimumPaneSize)
                 else: splitter.Unsplit(p1)