+++ src/sqlitely/database.py
@@ -8,7 +8,7 @@ Released under the MIT License.
 
 @author      Erki Suurjaak
 @created     21.08.2019
-@modified    15.09.2024
+@modified    01.10.2024
 ------------------------------------------------------------------------------
 """
 from collections import defaultdict, OrderedDict
@@ -21,6 +21,7 @@ import math
 import os
 import re
 import sqlite3
+import string
 import sys
 import tempfile
 
@@ -57,9 +58,29 @@ class Database(object):
     DATA_CATEGORIES = ["table", "view"]
 
     """SQLite features and the runtime library version they appeared in."""
-    FEATURE_SUPPORT = {"full_rename_table": (3, 25), "generated_column": (3, 31),
-                       "rename_column":     (3, 25), "strict":           (3, 37),
-                       "view_columns":      (3,  9)}
+    FEATURE_SUPPORT = {
+        "full_rename_table": (3, 25),
+        "generated_column":  (3, 31),
+        "rename_column":     (3, 25),
+        "strict":            (3, 37),
+        "view_columns":      (3,  9),
+    }
+
+    """SQLite PRAGMA extensions and the runtime library version they appeared in."""
+    PRAGMA_SUPPORT = {
+        "analysis_limit":      (3, 32),
+        "function_list":       (3, 30),
+        "hard_heap_limit":     (3, 31),
+        "legacy_alter_table":  (3, 26),
+        "module_list":         (3, 30),
+        "pragma_list":         (3, 30),
+        "trusted_schema":      (3, 31),
+    }
+
+    """SQLite PRAGMA extensions and the runtime library version they were deprecated in."""
+    PRAGMA_SUPPORT_DEPRECATED = {
+        "case_sensitive_like": (3, 44),
+    }
 
     """
     SQLite PRAGMA settings, as {
@@ -71,8 +92,8 @@ class Database(object):
         ?values:      {primitive: label},
         ?default:     directive default value,
         ?deprecated:  whether directive is deprecated,
-        ?dump:        whether directive should be included in db dump
-                      and statistics export,
+        ?dump:        whether directive should be included in db dump and statistics export,
+        ?format:      callable(row) for pragmas of type "table", returning row as formatted value,
         ?stats:       whether directive should be included in statistics export,
         ?initial:     whether directive should be issued before creating schema
                       or a callable(db, value) returning whether,
@@ -81,10 +102,20 @@ class Database(object):
         ?read:        false if setting is write-only,
         ?write:       false if setting is read-only
                       or a callable(db) returning false,
-        ?col:         result column to select if type "table"
     }.
     """
     PRAGMA = {
+      "analysis_limit": {
+        "name": "analysis_limit",
+        "label": "Analysis limit",
+        "type": int,
+        "short": "Limit on the approximate ANALYZE setting",
+        "description": """The approximate number of rows examined in each index by the ANALYZE command.
+0: analysis limit is disabled and the ANALYZE command will examine all rows of each index.
+> 0: ANALYZE commands will stop analyzing each index after it has examined approximately N rows.
+
+Setting to 100 or 1000 allows the ANALYZE command to run very quickly, even on multi-gigabyte database files. """
+      },
       "application_id": {
         "name": "application_id",
         "label": "Application ID",
@@ -135,6 +166,7 @@ Must be turned on before any tables are created, not possible to change afterwar
         "label": "Case-sensitive LIKE",
         "type": bool,
         "read": False,
+        "deprecated": PRAGMA_SUPPORT_DEPRECATED.get("case_sensitive_like", (sys.maxsize, )) >= sqlite3.sqlite_version_info,
         "short": "Case sensitivity on LIKE operator",
         "description": "Toggles case sensitivity on LIKE operator.",
       },
@@ -165,7 +197,6 @@ Must be turned on before any tables are created, not possible to change afterwar
         "name": "collation_list",
         "label": "Collation list",
         "type": "table",
-        "col": "name",
         "write": False,
         "short": "Collating sequences for current session",
         "description": "A list of the collating sequences defined for the current database connection.",
@@ -174,7 +205,6 @@ Must be turned on before any tables are created, not possible to change afterwar
         "name": "compile_options",
         "label": "Compile options",
         "type": "table",
-        "col": "compile_option",
         "write": False,
         "short": "SQLite compile-time options",
         "description": "Compile-time options used when building current SQLite library.",
@@ -271,6 +301,25 @@ Must be turned on before any tables are created, not possible to change afterwar
         "short": "Use Full FSYNC",
         "description": "Determines whether or not the F_FULLFSYNC syncing method is used on systems that support it (Mac OS-X only).",
       },
+      "function_list": {
+        "name": "function_list",
+        "label": "Function list",
+        "type": "table",
+        "format": lambda v: "%s(%s)" % (
+            v["name"], ".." if v["narg"] < 0 else ", ".join(("XYZ" + string.ascii_uppercase)[:v["narg"]]),
+        ),
+        "write": False,
+        "short": "SQL functions known to the database connection",
+        "description": "A list of SQL functions known to the database connection. Each row describes a single calling signature for a single SQL function. Some SQL functions will have multiple rows in the result set if they can (for example) be invoked with a varying number of arguments or can accept text in various encodings.",
+      },
+      "hard_heap_limit": {
+        "name": "hard_heap_limit",
+        "label": "Hard heap limit",
+        "type": int,
+        "min": 1,
+        "short": "Hard limit on allocated heap memory",
+        "description": "Hard upper bound of N bytes on the amount of heap memory that will be allocated. This pragma can only lower the heap limit, never raise it."
+      },
       "ignore_check_constraints": {
         "name": "ignore_check_constraints",
         "label": "Ignore check constraints",
@@ -298,6 +347,13 @@ Must be turned on before any tables are created, not possible to change afterwar
         "short": "Journal size byte limit",
         "description": """  Byte limit on the size of rollback-journal and WAL files left in the file-system after transactions or checkpoints. Each time a transaction is committed or a WAL file resets, SQLite compares the size of the rollback journal file or WAL file left in the file-system to the size limit set by this pragma, and if the journal or WAL file is larger, it is truncated to the limit.
   A negative number implies no limit. To always truncate rollback journals and WAL files to their minimum size, set to zero.""",
+      },
+      "legacy_alter_table": {
+        "name": "legacy_alter_table",
+        "label": "Legacy alter table",
+        "type": bool,
+        "short": "ALTER TABLE RENAME behavior",
+        "description": "If enabled, ALTER TABLE RENAME command only rewrites the initial occurrence of the table name in its CREATE TABLE statement and in any associated CREATE INDEX and CREATE TRIGGER statements. If disabled, all references to the table anywhere in the schema are converted to the new name.",
       },
       "legacy_file_format": {
         "name": "legacy_file_format",
@@ -331,6 +387,14 @@ Must be turned on before any tables are created, not possible to change afterwar
         "description": """  The maximum number of bytes that are set aside for memory-mapped I/O on a single database. If zero, memory mapped I/O is disabled.
   If negative, the limit reverts to the default value determined by the most recent sqlite3_config(SQLITE_CONFIG_MMAP_SIZE), or to the compile time default determined by SQLITE_DEFAULT_MMAP_SIZE if no start-time limit has been set.""",
       },
+      "module_list": {
+        "name": "module_list",
+        "label": "Module list",
+        "type": "table",
+        "write": False,
+        "short": "Virtual table modules",
+        "description": "A list of virtual table modules registered with the database connection.",
+      },
       "page_count": {
         "name": "page_count",
         "label": "Page count",
@@ -349,6 +413,14 @@ Must be turned on before any tables are created, not possible to change afterwar
         "initial": True,
         "short": "Database page byte size",
         "description": "The page size of the database. Specifying a new size does not change the page size immediately. Instead, the new page size is remembered and is used to set the page size when the database is first created, if it does not already exist when the page_size pragma is issued, or at the next VACUUM command that is run on the same database connection while not in WAL mode.",
+      },
+      "pragma_list": {
+        "name": "pragma_list",
+        "label": "Pragma list",
+        "type": "table",
+        "write": False,
+        "short": "Known PRAGMA commands",
+        "description": "A list of PRAGMA commands known to the database connection.",
       },
       "query_only": {
         "name": "query_only",
@@ -396,6 +468,14 @@ Must be turned on before any tables are created, not possible to change afterwar
         "short": "Result columns omit table name prefix",
         "description": "Affects the way SQLite names columns of data returned by SELECT statements.",
       },
+      "soft_heap_limit": {
+        "name": "soft_heap_limit",
+        "label": "Soft heap limit",
+        "type": int,
+        "min": 1,
+        "short": "Soft limit on allocated heap memory",
+        "description": "Advisory upper bound of N bytes on the amount of heap memory that will be allocated. SQLite strives to keep heap memory utilization below the soft heap limit by reducing the number of pages held in the page cache as heap memory usages approaches the limit."
+      },
       "synchronous": {
         "name": "synchronous",
         "label": "Synchronous",
@@ -438,6 +518,13 @@ Must be turned on before any tables are created, not possible to change afterwar
         "short": "Number of auxiliary threads for prepared statements",
         "description": "The upper bound on the number of auxiliary threads that a prepared statement is allowed to launch to assist with a query.",
       },
+      "trusted_schema": {
+        "name": "trusted_schema",
+        "label": "Trusted schema",
+        "type": bool,
+        "short": "Whether non-audited SQL functions and virtual tables are allowed",
+        "description": "Per-connection setting that determines whether or not SQL functions and virtual tables that have not been security audited are allowed to be run by views, triggers, or in expressions of the schema such as CHECK constraints, DEFAULT clauses, generated columns, expression indexes, and/or partial indexes.",
+      },
       "user_version": {
         "name": "user_version",
         "label": "User version",
@@ -468,8 +555,8 @@ WARNING: misuse can easily result in a corrupt database file.""",
     EXTRA_PRAGMAS = [
         "database_list", "foreign_key_check", "foreign_key_list", "incremental_vacuum",
         "index_info", "index_list", "index_xinfo", "integrity_check", "optimize",
-        "quick_check", "read_uncommitted", "shrink_memory", "soft_heap_limit",
-        "table_info", "table_xinfo", "wal_checkpoint",
+        "quick_check", "read_uncommitted", "shrink_memory", "table_info", "table_xinfo",
+        "wal_checkpoint",
     ]
 
     """Temporary file name counter."""
@@ -1484,9 +1571,10 @@ WARNING: misuse can easily result in a corrupt database file.""",
 
 
     @classmethod
-    def has_feature(cls, name):
+    def has_support(cls, name):
         """Returns whether the current SQLite version supports given feature."""
-        version_min = cls.FEATURE_SUPPORT.get(name, (sys.maxsize, ))
+        versions = cls.PRAGMA_SUPPORT if name in cls.PRAGMA_SUPPORT else cls.FEATURE_SUPPORT
+        version_min = versions.get(name, (sys.maxsize, ))
         return sqlite3.sqlite_version_info >= version_min
 
 
@@ -1821,7 +1909,9 @@ WARNING: misuse can easily result in a corrupt database file.""",
             if not rows:
                 if not callable(opts["type"]): continue # for name, opts
                 value = opts["type"]()
-            elif "table" == opts["type"]: value = [next(iter(x.values())) for x in rows]
+            elif "table" == opts["type"]:
+                getter = opts.get("format") or (lambda x: next(iter(x.values())))
+                value = sorted(getter(x) for x in rows)
             else:
                 value = next(iter(rows[0].values()))
                 if callable(opts["type"]): value = opts["type"](value)
@@ -1905,7 +1995,7 @@ WARNING: misuse can easily result in a corrupt database file.""",
         else:
             resets  = defaultdict(dict) # {category: {name: SQL}}
             if "table" == category \
-            and (name2 == grammar.quote(name2) or not self.has_feature("full_rename_table")):
+            and (name2 == grammar.quote(name2) or not self.has_support("full_rename_table")):
                 # Modify sqlite_master directly, as "ALTER TABLE x RENAME TO y"
                 # sets a quoted name "y" to CREATE statements, including related objects,
                 # regardless of whether the name required quoting.
@@ -1929,7 +2019,7 @@ WARNING: misuse can easily result in a corrupt database file.""",
                    for c in (item or {}).get("meta", {}).get("columns", [])):
             return
         table = item["name"]
-        if self.has_feature("rename_column"):
+        if self.has_support("rename_column"):
             altersql, err = grammar.generate(dict(
                 name=table, name2=table, columns=[(name, name2)]
             ), category="ALTER TABLE")
@@ -2243,8 +2333,12 @@ WARNING: misuse can easily result in a corrupt database file.""",
             logger.warning("Error syncing sqlite_master contents.", exc_info=True)
             try: self.execute("ROLLBACK")
             except Exception: pass
-if not Database.has_feature("strict"):
+
+if not Database.has_support("strict"):
     Database.AFFINITY.pop("ANY", None)
+for name in Database.PRAGMA_SUPPORT:
+    if not Database.has_support(name):
+        Database.PRAGMA.pop(name, None)